Podcast_AI_Project/requirements.txt
numpy
soundfile
scipy
noisereduce
openai-whisper
streamlit
wordcloud
matplotlib
sentence-transformers
transformers
nltk
torch
keybert
textblob
